requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
pandas
nltk
twilio
python-dotenv
python-telegram-bot
